+++ Packages/com.unity.cloud.gltfast.tests/Tests/Runtime/Scripts/JobTests.cs
@@ -1165,6 +1165,7 @@ namespace GLTFast.Tests.Jobs
         NativeArray<byte> m_InputUInt8;
         NativeArray<ushort> m_InputUInt16;
         NativeArray<uint> m_InputUInt32;
+        NativeArray<int> m_IndexInt32;
         NativeArray<int> m_IndexOutput;
 
         [OneTimeSetUp]
@@ -1173,13 +1174,15 @@ namespace GLTFast.Tests.Jobs
             m_InputUInt8 = new NativeArray<byte>(k_IndexLength, Allocator.Persistent);
             m_InputUInt16 = new NativeArray<ushort>(k_IndexLength, Allocator.Persistent);
             m_InputUInt32 = new NativeArray<uint>(k_IndexLength, Allocator.Persistent);
+            m_IndexInt32 = new NativeArray<int>(k_IndexLength, Allocator.Persistent);
             m_IndexOutput = new NativeArray<int>(k_IndexLength, Allocator.Persistent);
 
-            for (var i = 0; i < 6; i++)
+            for (var i = 0; i < k_IndexLength; i++)
             {
                 m_InputUInt8[i] = (byte)i;
                 m_InputUInt16[i] = (ushort)i;
                 m_InputUInt32[i] = (uint)i;
+                m_IndexInt32[i] = i;
             }
         }
 
@@ -1190,6 +1193,7 @@ namespace GLTFast.Tests.Jobs
             m_InputUInt16.Dispose();
             m_InputUInt32.Dispose();
             m_IndexOutput.Dispose();
+            m_IndexInt32.Dispose();
         }
 
         void CheckResult()
@@ -1208,6 +1212,32 @@ namespace GLTFast.Tests.Jobs
             Assert.AreEqual(3, m_IndexOutput[3]);
             Assert.AreEqual(5, m_IndexOutput[4]);
             Assert.AreEqual(4, m_IndexOutput[5]);
+        }
+
+        void CheckResultTriangleFan()
+        {
+            Assert.AreEqual(2, m_IndexOutput[0]);
+            Assert.AreEqual(1, m_IndexOutput[1]);
+            Assert.AreEqual(0, m_IndexOutput[2]);
+            Assert.AreEqual(3, m_IndexOutput[3]);
+            Assert.AreEqual(2, m_IndexOutput[4]);
+            Assert.AreEqual(0, m_IndexOutput[5]);
+            Assert.AreEqual(4, m_IndexOutput[6]);
+            Assert.AreEqual(3, m_IndexOutput[7]);
+            Assert.AreEqual(0, m_IndexOutput[8]);
+        }
+
+        void CheckResultTriangleStrip()
+        {
+            Assert.AreEqual(1, m_IndexOutput[0]);
+            Assert.AreEqual(0, m_IndexOutput[1]);
+            Assert.AreEqual(2, m_IndexOutput[2]);
+            Assert.AreEqual(3, m_IndexOutput[3]);
+            Assert.AreEqual(1, m_IndexOutput[4]);
+            Assert.AreEqual(2, m_IndexOutput[5]);
+            Assert.AreEqual(3, m_IndexOutput[6]);
+            Assert.AreEqual(2, m_IndexOutput[7]);
+            Assert.AreEqual(4, m_IndexOutput[8]);
         }
 
         [Test]
@@ -1237,6 +1267,30 @@ namespace GLTFast.Tests.Jobs
             Assert.AreEqual(5, m_IndexOutput[3]);
             Assert.AreEqual(4, m_IndexOutput[4]);
             Assert.AreEqual(3, m_IndexOutput[5]);
+        }
+
+        [Test]
+        public unsafe void CreateIndicesForTriangleFanJob()
+        {
+            Assert.IsTrue(m_IndexOutput.Length > 3);
+            var job = new GLTFast.Jobs.CreateIndicesForTriangleFanJob
+            {
+                result = (int*)m_IndexOutput.GetUnsafePtr()
+            };
+            job.Run(m_IndexOutput.Length);
+            CheckResultTriangleFan();
+        }
+
+        [Test]
+        public unsafe void CreateIndicesForTriangleStripJob()
+        {
+            Assert.IsTrue(m_IndexOutput.Length > 3);
+            var job = new GLTFast.Jobs.CreateIndicesForTriangleStripJob
+            {
+                result = (int*)m_IndexOutput.GetUnsafePtr()
+            };
+            job.Run(m_IndexOutput.Length);
+            CheckResultTriangleStrip();
         }
 
         [Test]
@@ -1315,6 +1369,34 @@ namespace GLTFast.Tests.Jobs
             };
             job.Run(m_IndexOutput.Length / 3);
             CheckResultFlipped();
+        }
+
+        [Test]
+        public unsafe void RecalculateIndicesForTriangleFanJob()
+        {
+            Assert.IsTrue(m_IndexOutput.Length > 3);
+            var job = new GLTFast.Jobs.RecalculateIndicesForTriangleFanJob
+            {
+                input = (int*)m_IndexInt32.GetUnsafeReadOnlyPtr(),
+                result = (int*)m_IndexOutput.GetUnsafePtr()
+            };
+            var triangleCount = m_IndexOutput.Length / 3;
+            job.Run(triangleCount);
+            CheckResultTriangleFan();
+        }
+
+        [Test]
+        public unsafe void RecalculateIndicesForTriangleStripJob()
+        {
+            Assert.IsTrue(m_IndexOutput.Length > 3);
+            var job = new GLTFast.Jobs.RecalculateIndicesForTriangleStripJob
+            {
+                input = (int*)m_IndexInt32.GetUnsafeReadOnlyPtr(),
+                result = (int*)m_IndexOutput.GetUnsafePtr()
+            };
+            var triangleCount = m_IndexOutput.Length / 3;
+            job.Run(triangleCount);
+            CheckResultTriangleStrip();
         }
     }