+++ Packages/com.unity.cloud.gltfast/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - (Test) `OpenGltfScene` with open glTF file dialog for convenient testing.
+- (Test) Tests for C# jobs that calculate or re-order indices.
 
 ### Changed
 - Node name is assigned earlier during instantiation, enabling easier node identification by name (partially fixes [#724](https://github.com/atteneder/glTFast/issues/724)).